moved character configurations to config file
