+++ server/src/content/student/StudentUseCases.ts
@@ -1,8 +1,6 @@
 import type { CharWithUseCases } from '../types'
 
-import { Student } from './Student'
-import { OnlineStore } from './useCases/OnlineStore'
-import { StudyRoom } from './useCases/StudyRoom'
+import { Student, OnlineStore, StudyRoom } from '../../../../config/student'
 
 export const StudentUseCases: CharWithUseCases = {
   characterId: Student.id,
